If a fork fails at the right time, "running" may be set when pid=-1. This can be very bad when the cleanup routine runs kill(pid, SIGKILL).

+++ spamass-milter.cpp
@@ -1,6 +1,6 @@
 // 
 //
-//  $Id: spamass-milter.cpp,v 1.82 2004/08/26 18:01:35 dnelson Exp $
+//  $Id: spamass-milter.cpp,v 1.83 2004/09/21 20:49:18 dnelson Exp $
 //
 //  SpamAss-Milter 
 //    - a rather trivial SpamAssassin Sendmail Milter plugin
@@ -127,7 +127,7 @@ int daemon(int nochdir, int noclose);
 
 // }}} 
 
-static const char Id[] = "$Id: spamass-milter.cpp,v 1.82 2004/08/26 18:01:35 dnelson Exp $";
+static const char Id[] = "$Id: spamass-milter.cpp,v 1.83 2004/09/21 20:49:18 dnelson Exp $";
 
 struct smfiDesc smfilter =
   {
@@ -1217,6 +1217,7 @@ mlfi_abort(SMFICTX* ctx)
 SpamAssassin::SpamAssassin():
   error(false),
   connected(false),
+  running(false),
   _numrcpt(0)
 {
 }
@@ -1235,12 +1236,15 @@ SpamAssassin::~SpamAssassin()
 		// child still running?
 		if (running)
 		{
-			// slaughter child
-			kill(pid, SIGKILL);
+			// make sure the pid is valid
+			if (pid > 0) {
+				// slaughter child
+				kill(pid, SIGKILL);
 
-			// wait for child to terminate
-			int status;
-			waitpid(pid, &status, 0);
+				// wait for child to terminate
+				int status;
+				waitpid(pid, &status, 0);
+			}
 		}
     }
